feat: revert restructure array summary
