+++ app/components/arraySummary/ArraySummaryDataItems.tsx
